Add test for utils and geometry

+++ Dissects/utils/utils.py
@@ -9,4 +9,4 @@ def pixel_to_um(df, pixel_size, coords, n_coords):
 
 def um_to_pixel(df, pixel_size, coords, n_coords):
     for i in range(len(coords)):
-    	df[n_coords[i]] = df[coords[i]]/pixel_size[coords[i]]
+    	df[n_coords[i]] = (df[coords[i]]/pixel_size[coords[i]]).astype(int)
